separated lrs and adl_lrs urls
