+++ lrs/templates/base.html
@@ -27,18 +27,18 @@
     <div class="home-menu pure-menu pure-menu-horizontal pure-menu-fixed">
         <div class="pure-g">
           <div class="pure-u-1-2">
-            <a class="pure-menu-heading" href="{% url lrs.views.home %}">ADL LRS</a>
+            <a class="pure-menu-heading" href="{% url adl_lrs.views.home %}">ADL LRS</a>
           </div>
           <div class="pure-u-1-2">
             <ul class="pure-menu-list">
                 {% if user.is_authenticated %}
-                  <li class="pure-menu-item"><a class="pure-menu-link" href="{% url lrs.views.me %}">My Account</a></li>
+                  <li class="pure-menu-item"><a class="pure-menu-link" href="{% url adl_lrs.views.me %}">My Account</a></li>
                   <li class="pure-menu-item"><a class="pure-menu-link" href="{% url logout %}">Logout</a></li>
                 {% else %}
                   <li class="pure-menu-item pure-menu-has-children pure-menu-allow-hover">
                     <a href="{% url login %}" id="menuLink1" class="pure-menu-link">Login</a>
                     <ul class="pure-menu-children">
-                        <li class="pure-menu-item"><a href="{% url lrs.views.register %}" class="pure-menu-link">Register</a></li>
+                        <li class="pure-menu-item"><a href="{% url adl_lrs.views.register %}" class="pure-menu-link">Register</a></li>
                     </ul>
                   </li>
                 {% endif %}
