+++ lrs/templates/home.html
@@ -90,7 +90,7 @@
           Want to make sure your statement is valid? Validate it here so you're not sending bad statements!
         </p>
         <p>
-          <a class="pure-button pull-right" href="{% url lrs.views.stmt_validator %}">Validate</a>
+          <a class="pure-button pull-right" href="{% url adl_lrs.views.stmt_validator %}">Validate</a>
         </p>
       </div>
       <div class="l-box pure-u-1 pure-u-md-1-2 pure-u-lg-1-5">
@@ -126,7 +126,7 @@
             <li>Once the client is authenticated, send the appropriate OAuth headers and values from your token to gain an access token.({% url oauth:oauth_provider.views.access_token %})</li>
             <li>You can now send this access token in with the OAuth headers when making calls to the LRS system</li>
           </ul>
-          <a class="pure-button" href="{% url lrs.views.reg_client %}">Register OAuth1 Client</a>
+          <a class="pure-button" href="{% url adl_lrs.views.reg_client %}">Register OAuth1 Client</a>
         </p>
     </div>
     <div class="l-box-lrg is-center pure-u-1 pure-u-md-1-2 pure-u-lg-2-5">
@@ -144,7 +144,7 @@
             </li>
             <li>You can now send this access token in with the OAuth2 headers when making calls to the LRS system</li>
           </ul>
-          <a class="pure-button" href="{% url lrs.views.reg_client2 %}">Register OAuth2 Client</a>
+          <a class="pure-button" href="{% url adl_lrs.views.reg_client2 %}">Register OAuth2 Client</a>
         </p>        
     </div>
   </div>
@@ -152,7 +152,7 @@
     <h2 class="content-head is-center">Sign up for HTTP Basic Credentials</h2>
       <div class="pure-g">
           <div class="l-box-lrg pure-u-1 pure-u-md-2-5">
-              <form class="pure-form pure-form-stacked" action="{% url lrs.views.register %}" method="post">
+              <form class="pure-form pure-form-stacked" action="{% url adl_lrs.views.register %}" method="post">
                   {% csrf_token %}
                   <fieldset>
                       {% for field in form %}
