+++ lrs/templates/me.html
@@ -13,8 +13,8 @@
             <div class="pure-menu custom-restricted-width custom-menu">
                 <ul class="pure-menu-list">
                     <li class="pure-menu-item pure-menu-item-custom pure-menu-selected tab-content"><a href="#myaccount" id="showme" class="pure-menu-link custom-wrap">Account</a></li>
-                    <li class="pure-menu-item pure-menu-item-custom"><a target="_blank" href="{% url lrs.views.my_statements %}" class="pure-menu-link custom-wrap">Statements</a></li>
-                    <li class="pure-menu-item pure-menu-item-custom"><a target="_blank" href="{% url lrs.views.my_activity_states %}" class="pure-menu-link custom-wrap">Activity States</a></li>
+                    <li class="pure-menu-item pure-menu-item-custom"><a target="_blank" href="{% url adl_lrs.views.my_statements %}" class="pure-menu-link custom-wrap">Statements</a></li>
+                    <li class="pure-menu-item pure-menu-item-custom"><a target="_blank" href="{% url adl_lrs.views.my_activity_states %}" class="pure-menu-link custom-wrap">Activity States</a></li>
                     <li class="pure-menu-item pure-menu-item-custom tab-content"><a href="#clientapps" class="pure-menu-link custom-wrap">OAuth Client Apps</a></li>
                     <li class="pure-menu-item pure-menu-item-custom tab-content"><a href="#myaccesstokens" class="pure-menu-link custom-wrap">OAuth Access Tokens</a></li>
                     <li class="pure-menu-item pure-menu-item-custom tab-content"><a href="#clientapps2" class="pure-menu-link custom-wrap">OAuth2 Client Apps</a></li>
@@ -44,7 +44,7 @@
             </div>
             <div id="clientapps" class="content">
                 <h2>OAuth Client Apps</h2>
-                <a class="pure-button" href="{% url lrs.views.reg_client %}">Add</a>
+                <a class="pure-button" href="{% url adl_lrs.views.reg_client %}">Add</a>
                 {% if client_apps %}
                     <div class="email-item">
                         {% for app in client_apps %}
@@ -90,7 +90,7 @@
             </div>
             <div id="clientapps2" class="content">
                 <h2>OAuth2 Client Apps</h2>
-                <a class="pure-button" href="{% url lrs.views.reg_client2 %}">Add</a>
+                <a class="pure-button" href="{% url adl_lrs.views.reg_client2 %}">Add</a>
                 {% if client_apps2 %}
                     <div class="email-item">
                         {% for app in client_apps2 %}
@@ -160,7 +160,7 @@
                 status = "Accepted";                
             } 
             $.ajax({
-                url: "{% url lrs.views.my_app_status %}?app_name="+appname+"&status="+status,
+                url: "{% url adl_lrs.views.my_app_status %}?app_name="+appname+"&status="+status,
                 type: "POST",
                 context: $(this),
                 success: function (data){
@@ -187,7 +187,7 @@
         $('.delete_token').click(function(){
             var id = $(this).next('span').text();
             $.ajax({
-                url: "{% url lrs.views.delete_token %}?id="+id,
+                url: "{% url adl_lrs.views.delete_token %}?id="+id,
                 type: "DELETE",
                 context: $(this),
                 success: function(data){
@@ -202,7 +202,7 @@
         $('.delete_token2').click(function(){
             var id = $(this).next('span').text();
             $.ajax({
-                url: "{% url lrs.views.delete_token2 %}?id="+id,
+                url: "{% url adl_lrs.views.delete_token2 %}?id="+id,
                 type: "DELETE",
                 context: $(this),
                 success: function(data){
@@ -217,7 +217,7 @@
         $('.delete_client').click(function(){
             var id = $(this).next('span').text();
             $.ajax({
-                url: "{% url lrs.views.delete_client %}?id="+id,
+                url: "{% url adl_lrs.views.delete_client %}?id="+id,
                 type: "DELETE",
                 context: $(this),
                 success: function(data){
