+++ lrs/templates/regclient2.html
@@ -17,7 +17,7 @@
 	        {% if form.non_field_errors %}
 	            {{ form.non_field_errors }}
 	        {% endif %}
-		    <form class="pure-form pure-form-stacked" action="{% url lrs.views.reg_client2 %}" method="post">
+		    <form class="pure-form pure-form-stacked" action="{% url adl_lrs.views.reg_client2 %}" method="post">
 		      {% csrf_token %}
 		      <fieldset>
 		          {% for field in form %}
