+++ lrs/templates/validator.html
@@ -23,7 +23,7 @@
 	            <li>{{ valid_message }}</li>
 	          </ul>
 			{% endif %}
-			<form class="pure-form pure-form-stacked" action="{% url lrs.views.stmt_validator %}" method="post">
+			<form class="pure-form pure-form-stacked" action="{% url adl_lrs.views.stmt_validator %}" method="post">
 				{% csrf_token %}
                   <fieldset>
                       {% for field in form %}
