+++ lrs/tests/AgentProfileTests.py
@@ -8,7 +8,8 @@ from django.test import TestCase
 from django.conf import settings
 from django.core.urlresolvers import reverse
 
-from ..views import register, agent_profile
+from ..views import agent_profile
+from adl_lrs.views import register
 
 class AgentProfileTests(TestCase):
     testagent = '{"mbox":"mailto:test@example.com"}'
