+++ lrs/tests/StatementFilterTests.py
@@ -18,8 +18,9 @@ from django.core.urlresolvers import reverse
 from django.conf import settings
 
 from ..models import Statement
-from ..views import register, statements, statements_more
+from ..views import statements, statements_more
 from ..util.util import convert_to_utc
+from adl_lrs.views import register
 
 class StatementFilterTests(TestCase):
